FIXED leoz-ui: command line interface breakage on linux

+++ leoz-ui/src/main/java/org/deku/leoz/ui/Setup.kt
@@ -1,6 +1,7 @@
 package org.deku.leoz.ui
 
 import com.sun.jna.platform.win32.Shell32Util
+import org.apache.commons.lang3.SystemUtils
 import org.deku.leoz.bundle.Bundle
 import org.deku.leoz.bundle.BundleProcessInterface
 import org.slf4j.LoggerFactory
@@ -17,20 +18,37 @@ class Setup : BundleProcessInterface() {
     private val bundle by lazy({
         Bundle.load(this.javaClass)
     })
-    private val desktopLinkFile  by lazy({
-        val desktopFolderPath = Shell32Util.getSpecialFolderPath(CSIDL_DESKTOP, false)
-        File(desktopFolderPath).resolve("${this.bundle.name!!.capitalize()}.lnk")
-    })
+    private val desktopLinkFile by lazy {
+        when {
+            SystemUtils.IS_OS_WINDOWS -> {
+                val desktopFolderPath = Shell32Util.getSpecialFolderPath(CSIDL_DESKTOP, false)
+                File(desktopFolderPath).resolve("${this.bundle.name!!.capitalize()}.lnk")
+            }
+            else -> {
+                null
+            }
+        }
+    }
 
     override fun install() {
-        val shellLink = WindowsShellLink(this.desktopLinkFile.absolutePath)
-        shellLink.target = bundle.path!!.resolve("${bundle.name!!}.exe").absolutePath
-        shellLink.save()
+        when {
+            SystemUtils.IS_OS_WINDOWS -> {
+                val desktopLinkFile = this.desktopLinkFile
+                if (desktopLinkFile != null) {
+                    val shellLink = WindowsShellLink(desktopLinkFile.absolutePath)
+                    shellLink.target = bundle.path!!.resolve("${bundle.name!!}.exe").absolutePath
+                    shellLink.save()
+                }
+            }
+        }
     }
 
     override fun uninstall() {
-        if (desktopLinkFile.exists())
-            desktopLinkFile.delete()
+        val desktopLinkFile = this.desktopLinkFile
+        if (desktopLinkFile != null) {
+            if (desktopLinkFile.exists())
+                desktopLinkFile.delete()
+        }
     }
 
     override fun start() {
